Don't reset lastOpReceivedAt on worker iteration #49
If a socket read times out the worker loop restarts. Restarting the worker loop causes
lastOpReceivedAt to be reset. If the socket reads continually time out because no data
has been received for some time, the tests for ConsumerMaxMsSilenceFromServer and
ConsumerPingAfterMsSilenceFromServer will never be met.

This moves the initial assignment to lastOpReceivedAt to outside the worker's main
loop, which enables the server silence calculations to succeed.